updated meet the team page
