+++ frontend/app/globals.css
@@ -10,4 +10,22 @@
     font-weight: 400;
     font-style: normal;
   }
+  
+@font-face {
+    font-family: 'Light';
+    src: url('/Satoshi-Variable.ttf') format('truetype');
+    font-weight: 300;
+    font-style: normal;
+  }
 
+  .fade-slide-up {
+    opacity: 0;
+    transform: translateY(20px);
+    transition: opacity 0.5s ease-out, transform 0.5s ease-out;
+  }
+  
+  .fade-slide-up.visible {
+    opacity: 1;
+    transform: translateY(0);
+  }
+
